Added single and local multiplayer menus
Created two separate game setup menus for these modes, with the single
player mode having options for easy, medium and hard difficulty, while
the multiplayer mode has an option to turn the timer bar off. Also
switched the position of the time selection and the number of games to
make turning the timer off look nicer (the time selection disappears).